docker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
InstructorEmbedding==1.0.1
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
transformers==4.36.2
sentence-transformers==2.3.1
huggingface-hub==0.20.3
